Implement authentication routes and middleware, update user model for citizenid type, and configure server settings

+++ models/user.js
@@ -19,7 +19,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   user.init({
     emid: DataTypes.INTEGER,
-    citizenid: DataTypes.INTEGER,
+    citizenid: DataTypes.BIGINT,
     password: DataTypes.STRING,
     prefixID: DataTypes.INTEGER,
     fname: DataTypes.STRING,
@@ -31,6 +31,6 @@ module.exports = (sequelize, DataTypes) => {
   }, {
     sequelize,
     modelName: 'user',
-  });
+  }); 
   return user;
 };
